+++ documentation/changelog.md
@@ -24,6 +24,9 @@ description: List of all the changes and current versions.
 + Added Documentation for Client Plugins
 + Added Custom Job Preview Panel
 + Added /quitjob
++ Added PgUp & PgDn for Chat
++ Fixed Clothing Menu Drawables / Textures Overflow
++ Fixed Design Choices for Clothing Menu
 ```
 
 ## 1.6.0 - February 1-10, 2021